tests/wpt/web-platform-tests/user-timing/mark-errors.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>UserTimingL3: errors are thrown when mark() is called incorrectly.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<h1>UserTimingL3: Mark</h1>
<script>
  test(function() {
    assert_throws(new TypeError(), function() { self.performance.mark("mark1", 123); }, "Number passed as a dict argument should cause type-error.")
  }, "Number should be rejected as the mark-options.")

  test(function() {
    assert_throws(new TypeError(), function() { self.performance.mark("mark1", NaN); }, "NaN passed as a dict argument should cause type-error.")
  }, "NaN should be rejected as the mark-options.")

  test(function() {
    assert_throws(new TypeError(), function() { self.performance.mark("mark1", Infinity); }, "Infinity passed as a dict argument should cause type-error.")
  }, "Infinity should be rejected as the mark-options.")

  test(function() {
    assert_throws(new TypeError(), function() { self.performance.mark("mark1", "string"); }, "String passed as a dict argument should cause type-error.")
  }, "String should be rejected as the mark-options.")
</script>
